Report System

+++ events/AntiScam.js
@@ -1,23 +1,24 @@
-const { Events, EmbedBuilder } = require("discord.js");
+const { Events } = require("discord.js");
 
 module.exports = {
     name: Events.MessageCreate,
     async execute(message, client) {
-        if (message.channelId != '1213789140555137025') return
+        if (message.channelId != '1213789140555137025') return;
+        
+        const guild = message.guild;
+
+        // const adminRole = await guild.roles.cache.get("1213411732391723068");
+        // const modRole = await guild.roles.cache.get("1213126421875269652");
+        // const staffRole = await guild.roles.cache.get("1213126482730291291");
+
+        const hasVerify = await message.member.roles.cache.has("1213411732391723068");
+        const hasMod = await message.member.roles.cache.has("1213126421875269652");
+        const hasStaff = await message.member.roles.cache.has("1213126482730291291");
+
+        if (hasVerify || hasMod || hasStaff) return;
 
         await message.delete();
 
-        await message.member.send({
-            embeds: [
-                new EmbedBuilder()
-                .setColor('Red')
-                .setDescription('⛔ คุณถูกแบนเนื่องจากสงสัยว่าเป็นสแปม\nหากคิดว่าเป็นเรื่องเข้าใจผิดกรุณาส่งเรื่องไปยัง **[Dek Pua Support](https://www.dekpua.com/support?action=ticket)**.\n\n⛔ You have been banned due to suspicion of spam.\nIf you believe this is a misunderstanding, please submit a ticket to **[Dek Pua Support](https://www.dekpua.com/support?action=ticket)**.')
-            ]
-        });
-
-        await message.member.ban({
-            deleteMessageSeconds: 60 * 60 * 24,
-            reason: "[AntiScam] สงสัยว่าเป็นสแปม Suspected of spam"
-        });
+        await client.ban(message.member, { th: "[AntiScam] สงสัยว่าเป็นสแปม", en: "[AntiScam] Suspected of spam"})
     }
 }
